requirements.txt
requests
beautifulsoup4
pandas
matplotlib
torch
transformers
fastapi
uvicorn
google-generativeai
emoji
unidecode
langchain